+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support `patternProperties` when converting to a dynamic object
+
 ### Fixed
 - Fix some anomaly checking with `patternProperties`
 - Fix some anomaly checking with `additionalProperties`